refactor(Core): create app state
