+++ src/app/core/auth/auth.actions.ts
@@ -1,0 +1,16 @@
+import { Action } from '@ngrx/store';
+
+export enum AuthActionTypes {
+  LOGIN = '[Auth] Login',
+  LOGOUT = '[Auth] Logout'
+}
+
+export class ActionAuthLogin implements Action {
+  readonly type = AuthActionTypes.LOGIN;
+}
+
+export class ActionAuthLogout implements Action {
+  readonly type = AuthActionTypes.LOGOUT;
+}
+
+export type AuthActions = ActionAuthLogin | ActionAuthLogout;
