+++ src/app/core/auth/auth.models.ts
@@ -1,0 +1,3 @@
+export interface AuthState {
+  isAuthenticated: boolean;
+}
