+++ src/app/core/auth/auth.selectors.ts
@@ -1,0 +1,8 @@
+import { createSelector } from '@ngrx/store';
+import { selectAuthState } from '../core.state';
+import { AuthState } from './auth.models';
+
+export const selectAuth = createSelector(
+  selectAuthState,
+  (state: AuthState) => state
+);
