+++ src/app/core/core.module.ts
@@ -1,31 +1,19 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { MetaReducer, StoreModule } from '@ngrx/store';
+import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
-import { storeFreeze } from 'ngrx-store-freeze';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
 import { environment } from '@env/environment';
 
-import { debug } from './meta-reducers/debug.reducer';
-import { initStateFromLocalStorage } from './meta-reducers/init-state-from-local-storage.reducer';
 import { LocalStorageService } from './local-storage/local-storage.service';
-import { authReducer } from './auth/auth.reducer';
 import { AuthEffects } from './auth/auth.effects';
 import { AuthGuardService } from './auth/auth-guard.service';
 import { AnimationsService } from './animations/animations.service';
 import { TitleService } from './title/title.service';
-
-export const metaReducers: MetaReducer<any>[] = [initStateFromLocalStorage];
-
-if (!environment.production) {
-  metaReducers.unshift(storeFreeze);
-  if (!environment.test) {
-    metaReducers.unshift(debug);
-  }
-}
+import { reducers, metaReducers } from './core.state';
 
 @NgModule({
   imports: [
@@ -34,12 +22,7 @@ if (!environment.production) {
     HttpClientModule,
 
     // ngrx
-    StoreModule.forRoot(
-      {
-        auth: authReducer
-      },
-      { metaReducers }
-    ),
+    StoreModule.forRoot(reducers, { metaReducers: metaReducers() }),
     EffectsModule.forRoot([AuthEffects]),
 
     // 3rd party
