+++ src/app/core/index.ts
@@ -2,6 +2,8 @@ export * from './local-storage/local-storage.service';
 export * from './animations/route.animations';
 export * from './animations/animations.service';
 export * from './auth/auth.reducer';
+export * from './auth/auth.actions';
+export * from './auth/auth.selectors';
 export * from './auth/auth-guard.service';
 export * from './title/title.service';
 export * from './core.module';
